move attribute_handling tests to their own module
